+++ invoice-tracker-backend/src/main/java/net/talaatharb/invoicetracker/dtos/UserDto.java
@@ -41,10 +41,10 @@ public class UserDto {
     private String arabicAddress;
 
     //		@NotBlank
-    private float allowedBalance;
+    private int allowedBalance;
 
     //	@NotBlank
-    private float remainingBalance;
+    private int remainingBalance;
 
     //	@NotBlank
     private boolean billable;
